[src] Update dataset_untrimmed module
Based on the updates proposed to the JSON and HDF5 files, we have to
update the dataset module accordingly.

- Update DiDeMo notebook to notify the removal of `num_clips` from the
  JSON.

- Minor refactoring to proposals module.

- Fix training loop due to sanity check added lately.

+++ train.py
@@ -33,7 +33,7 @@ METRICS_OLD = ['iou', 'r@1', 'r@5']
 VARS_TO_RECORD = ['id', 'hit@iou,k', 'scores', 'topk_segments']
 TRACK = 'r@1,0.7'
 BEST_RESULT = 0.0
-TOPK_DIDEMO = torch.tensor([1, 5])
+TOPK_DIDEMO = torch.tensor([1, 5]).float()
 
 parser = argparse.ArgumentParser(
     description='*MCN training',
